fix(helpers): remove deprecated helpers
BREAKING CHANGE: remove instantsearch.highlight, instantsearch.reverseHighlight, instantsearch.snippet, instantsearch.reverseSnippet. Instead use the `components` available from a template
BREAKING CHANGE: remove instantsearch.insights, instead use sendEvent
BREAKING CHANGE: remove instantsearch.getInsightsAnonymousUserToken
