+++ packages/instantsearch.js/src/connectors/index.ts
@@ -3,10 +3,8 @@ export { default as connectClearRefinements } from './clear-refinements/connectC
 export { default as connectCurrentRefinements } from './current-refinements/connectCurrentRefinements';
 export { default as connectHierarchicalMenu } from './hierarchical-menu/connectHierarchicalMenu';
 export { default as connectHits } from './hits/connectHits';
-export { default as connectHitsWithInsights } from './hits/connectHitsWithInsights';
 export { default as connectHitsPerPage } from './hits-per-page/connectHitsPerPage';
 export { default as connectInfiniteHits } from './infinite-hits/connectInfiniteHits';
-export { default as connectInfiniteHitsWithInsights } from './infinite-hits/connectInfiniteHitsWithInsights';
 export { default as connectMenu } from './menu/connectMenu';
 export { default as connectNumericMenu } from './numeric-menu/connectNumericMenu';
 export { default as connectPagination } from './pagination/connectPagination';
