--- packages/instantsearch.js/src/helpers/get-insights-anonymous-user-token.ts
@@ -1,39 +1,0 @@
-import { warning } from '../lib/utils';
-
-export const ANONYMOUS_TOKEN_COOKIE_KEY = '_ALGOLIA';
-
-function getCookie(name: string): string | undefined {
-  if (typeof document !== 'object' || typeof document.cookie !== 'string') {
-    return undefined;
-  }
-
-  const prefix = `${name}=`;
-  const cookies = document.cookie.split(';');
-  for (let i = 0; i < cookies.length; i++) {
-    let cookie = cookies[i];
-    while (cookie.charAt(0) === ' ') {
-      cookie = cookie.substring(1);
-    }
-    if (cookie.indexOf(prefix) === 0) {
-      return cookie.substring(prefix.length, cookie.length);
-    }
-  }
-  return undefined;
-}
-
-export function getInsightsAnonymousUserTokenInternal(): string | undefined {
-  return getCookie(ANONYMOUS_TOKEN_COOKIE_KEY);
-}
-
-/**
- * @deprecated This function will be still supported in 4.x releases, but not further. It is replaced by the `insights` middleware. For more information, visit https://www.algolia.com/doc/guides/getting-insights-and-analytics/search-analytics/click-through-and-conversions/how-to/send-click-and-conversion-events-with-instantsearch/js/
- */
-export default function getInsightsAnonymousUserToken(): string | undefined {
-  warning(
-    false,
-    `\`getInsightsAnonymousUserToken\` function has been deprecated. It is still supported in 4.x releases, but not further. It is replaced by the \`insights\` middleware.
-
-For more information, visit https://www.algolia.com/doc/guides/getting-insights-and-analytics/search-analytics/click-through-and-conversions/how-to/send-click-and-conversion-events-with-instantsearch/js/`
-  );
-  return getInsightsAnonymousUserTokenInternal();
-}
