--- packages/instantsearch.js/src/helpers/index.ts
@@ -1,13 +1,0 @@
-export * from './highlight';
-export * from './reverseHighlight';
-export * from './snippet';
-export * from './reverseSnippet';
-export { default as reverseHighlight } from './reverseHighlight';
-export { default as reverseSnippet } from './reverseSnippet';
-export { default as highlight } from './highlight';
-export { default as snippet } from './snippet';
-export { default as insights } from './insights';
-export {
-  default as getInsightsAnonymousUserToken,
-  getInsightsAnonymousUserTokenInternal,
-} from './get-insights-anonymous-user-token';
