--- packages/instantsearch.js/src/lib/insights/index.ts
@@ -1,5 +1,0 @@
-export {
-  default as withInsights,
-  inferPayload as inferInsightsPayload,
-} from './client';
-export { default as withInsightsListener } from './listener';
