+++ packages/vue-instantsearch/src/components/__tests__/Hits.js
@@ -73,65 +73,6 @@ it('exposes banner prop to the banner slot', () => {
   expect(img.attributes('src')).toBe('https://via.placeholder.com/550x250');
 });
 
-it('exposes insights prop to the default slot', async () => {
-  const insights = jest.fn();
-  __setState({
-    ...defaultState,
-    insights,
-  });
-
-  const wrapper = mount({
-    components: { Hits },
-    template: `
-      <Hits>
-        <template v-slot="{ items, insights }">
-          <ul>
-            <li v-for="(item, itemIndex) in items" >
-
-              <button :id="'add-to-cart-' + item.objectID" @click="insights('clickedObjectIDsAfterSearch', {eventName: 'Add to cart', objectIDs: [item.objectID]})">
-                Add to cart
-              </button>
-            </li>
-          </ul>
-        </template>
-      </Hits>
-    `,
-  });
-  await wrapper.find('#add-to-cart-two').trigger('click');
-  expect(insights).toHaveBeenCalledWith('clickedObjectIDsAfterSearch', {
-    eventName: 'Add to cart',
-    objectIDs: ['two'],
-  });
-});
-
-it('exposes insights prop to the item slot', async () => {
-  const insights = jest.fn();
-  __setState({
-    ...defaultState,
-    insights,
-  });
-
-  const wrapper = mount({
-    components: { Hits },
-    template: `
-      <Hits>
-        <template v-slot:item="{ item, insights }">
-          <div>
-            <button :id="'add-to-cart-' + item.objectID" @click="insights('clickedObjectIDsAfterSearch', {eventName: 'Add to cart', objectIDs: [item.objectID]})">
-              Add to cart
-            </button>
-          </div>
-        </template>
-      </Hits>
-    `,
-  });
-  await wrapper.find('#add-to-cart-two').trigger('click');
-  expect(insights).toHaveBeenCalledWith('clickedObjectIDsAfterSearch', {
-    eventName: 'Add to cart',
-    objectIDs: ['two'],
-  });
-});
-
 it('exposes send-event method for insights middleware', async () => {
   const sendEvent = jest.fn();
   __setState({
